Demo move DeployableReplicateToOmenAgent to examples
